requirements.txt
google-cloud-vision
uvicorn
fastapi
python-dotenv
google-generativeai
pydantic
pydantic_settings
python-multipart